requirements.txt
Django==4.0.4
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
whitenoise==6.0.0
# psycopg2-binary==2.9.3
